fix(ui): reconcile composer stack with current main
