+++ ui/src/pages/chat/chat-responsive.browser.test.ts
@@ -3159,7 +3159,7 @@ describeBrowserLayout.concurrent("chat responsive browser layout", () => {
     }
   });
 
-  it("keeps the expanded mobile composer tight, scrollable, and separated from the thread", async () => {
+  it("keeps the expanded mobile composer tight, scrollable, and flush with the thread", async () => {
     const page = await openFixture(393, 852);
     try {
       const textarea = page.locator(".agent-chat__composer-combobox > textarea");
@@ -3247,9 +3247,9 @@ describeBrowserLayout.concurrent("chat responsive browser layout", () => {
       // inside it, so a long draft can never push the thread off the page.
       expect(textareaRect.height).toBeLessThanOrEqual(layout.viewportHeight * 0.25 + 1);
       expect(textareaMetrics.scrollHeight).toBeGreaterThan(textareaMetrics.clientHeight);
-      expect(input.y - (thread.y + thread.height)).toBeGreaterThanOrEqual(5.5);
-      expect(shell.x).toBeLessThanOrEqual(8);
-      expect(layout.viewportWidth - (shell.x + shell.width)).toBeLessThanOrEqual(8);
+      expect(input.y - (thread.y + thread.height)).toBeCloseTo(0, 0);
+      expect(shell.x).toBeLessThanOrEqual(12);
+      expect(layout.viewportWidth - (shell.x + shell.width)).toBeLessThanOrEqual(12);
       expect(attach.x - input.x).toBeLessThanOrEqual(10);
       expect(model.x).toBeGreaterThanOrEqual(context.x + context.width - 1);
       expect(input.x + input.width - (send.x + send.width)).toBeLessThanOrEqual(8);
@@ -3588,7 +3588,7 @@ describeBrowserLayout.concurrent("chat responsive browser layout", () => {
     },
   );
 
-  it("preserves the mobile composer bottom spacing when its textarea is focused", async () => {
+  it("keeps the compact mobile composer bottom edge stable when its textarea is focused", async () => {
     const page = await openFixture(390, 844);
     try {
       const shell = page.locator(".agent-chat__composer-shell");
@@ -3599,7 +3599,7 @@ describeBrowserLayout.concurrent("chat responsive browser layout", () => {
       const focused = await readPosition();
 
       expect(focused).toBe(unfocused);
-      expect(focused).toBe("14px");
+      expect(focused).toBe("0px");
     } finally {
       await closeBrowserPage(page);
     }
