+++ ui/src/pages/new-session/composer-dictation-control.ts
@@ -36,10 +36,6 @@ export class NewSessionDictationControl {
     this.devicePicker = new ComposerMicrophonePicker(options.requestUpdate);
   }
 
-  get locked(): boolean {
-    return this.dictation?.locksComposer === true;
-  }
-
   dispose(): void {
     this.owner = null;
     this.dictation?.dispose();
