+++ ui/src/pages/new-session/draft-composer.ts
@@ -1,0 +1,113 @@
+import { nothing, type TemplateResult } from "lit";
+import type { GatewayAgentRow } from "../../api/types.ts";
+import type { ApplicationContext } from "../../app/context.ts";
+import type { ImageLightboxItem } from "../../components/image-lightbox.ts";
+import type { SessionToolOverrides } from "../../lib/sessions/patch.ts";
+import { refreshSlashCommands } from "../chat/chat-commands.ts";
+import type { CapabilityMenuProps } from "../chat/components/chat-composer-types.ts";
+import type { NewSessionAttachmentDraft } from "./attachment-draft.ts";
+import { NewSessionComposerTextareaController, renderNewSessionComposer } from "./composer.ts";
+import type { NewSessionVisibility } from "./create-params.ts";
+import type { NewSessionModelControl } from "./model-control.ts";
+
+export function renderNewSessionDraftComposer(options: {
+  agent?: GatewayAgentRow;
+  agentId: string;
+  attachmentDraft: NewSessionAttachmentDraft;
+  canSubmit: boolean;
+  context: ApplicationContext | undefined;
+  draftOwnerKey: string;
+  isCatalogTarget: boolean;
+  message: string;
+  visibility?: NewSessionVisibility;
+  draftAvailable?: boolean;
+  capabilityMenu?: CapabilityMenuProps;
+  toolOverrides?: SessionToolOverrides | null;
+  modelControl: NewSessionModelControl;
+  permissionControl?: TemplateResult;
+  textareaController: NewSessionComposerTextareaController;
+  voiceControl?: TemplateResult | typeof nothing;
+  requiresModifier: boolean;
+  requestUpdate: () => void;
+  submitDisabledReason?: string;
+  blockedSubmitNotice?: string;
+  dictationActive?: boolean;
+  dictationPreview?: string;
+  dictationStatus?: TemplateResult | typeof nothing;
+  terminalAction?: {
+    canStart: boolean;
+    disabledReason?: string;
+    onStart: () => void;
+  };
+  submitting: boolean;
+  messageLocked?: boolean;
+  onInput: (message: string) => void;
+  onOpenImage?: (item: ImageLightboxItem) => void;
+  onVisibilityChange?: (visibility: NewSessionVisibility) => void;
+  onSubmit: () => void;
+}) {
+  const readSignal = options.attachmentDraft.readSignal;
+  const commandClient = options.context?.gateway.snapshot.client ?? null;
+  options.textareaController.syncSkillCommandOwner(
+    commandClient,
+    options.agentId,
+    options.draftOwnerKey,
+  );
+  return renderNewSessionComposer({
+    attachmentLimits: options.context?.gateway.snapshot.hello?.policy?.attachments,
+    attachments: options.attachmentDraft.attachments,
+    canSubmit: options.canSubmit,
+    getAttachments: () => options.attachmentDraft.attachments,
+    message: options.message,
+    visibility: options.visibility,
+    draftAvailable: options.draftAvailable,
+    capabilityMenu: options.capabilityMenu,
+    toolOverrides: options.toolOverrides,
+    modelControl: options.isCatalogTarget
+      ? nothing
+      : options.modelControl.render({
+          agent: options.agent,
+          agentId: options.agentId,
+          context: options.context,
+          sending: options.submitting,
+        }),
+    permissionControl: options.permissionControl,
+    pendingAttachmentReads: options.attachmentDraft.pendingReads,
+    readSignal,
+    requiresModifier: options.requiresModifier,
+    requestUpdate: options.requestUpdate,
+    refreshCommands: commandClient
+      ? () =>
+          refreshSlashCommands({
+            client: commandClient,
+            agentId: options.agentId,
+            shouldApply: () =>
+              options.textareaController.ownsSkillCommands(
+                commandClient,
+                options.agentId,
+                options.draftOwnerKey,
+              ),
+          })
+      : undefined,
+    submitDisabledReason: options.submitDisabledReason,
+    blockedSubmitNotice: options.blockedSubmitNotice,
+    dictationActive: options.dictationActive,
+    dictationPreview: options.dictationPreview,
+    dictationStatus: options.dictationStatus,
+    terminalAction: options.terminalAction,
+    submitting: options.submitting,
+    textareaController: options.textareaController,
+    voiceControl: options.voiceControl,
+    messageLocked: options.messageLocked,
+    onAttachmentsChange: (attachments) => {
+      if (!options.submitting && !options.messageLocked) {
+        options.attachmentDraft.replace(attachments);
+      }
+    },
+    onPendingReadsChange: (delta) => options.attachmentDraft.updatePending(readSignal, delta),
+    onInput: options.onInput,
+    onOpenImage: options.onOpenImage,
+    onVisibilityChange: options.onVisibilityChange,
+    onSubmit: options.onSubmit,
+  });
+}
