+++ ui/src/pages/new-session/draft-submission-flow.ts
@@ -59,7 +59,7 @@ export class DraftSubmissionFlow {
   private messageValue = "";
   private submittingValue = false;
   private blockedSubmitGate: string | null = null;
-  private submissionOutcomeUnknownValue: SubmissionOutcomeReason | null = null;
+  submissionOutcomeUnknown: SubmissionOutcomeReason | null = null;
   private readonly startedSession = new StartedSessionNavigation();
   error: string | null = null;
   private submitRequestToken = 0;
@@ -117,16 +117,12 @@ export class DraftSubmissionFlow {
     return this.submittingValue || this.sessionStartup.active;
   }
 
-  get submissionOutcomeUnknown(): SubmissionOutcomeReason | null {
-    return this.submissionOutcomeUnknownValue;
-  }
-
   resumeInterruptedSubmission() {
     const startup = this.sessionStartup.resume();
     if (startup.kind === "resume") {
       void this.submit(startup);
     } else if (startup.kind !== "wait") {
-      this.submissionOutcomeUnknownValue = "gateway-changed";
+      this.submissionOutcomeUnknown = "gateway-changed";
       this.callbacks.requestUpdate();
     }
   }
@@ -191,7 +187,7 @@ export class DraftSubmissionFlow {
 
   markPendingPlacementUnavailable(outcome: SubmissionOutcomeReason) {
     this.pendingPlacement.retryAllowed = false;
-    this.submissionOutcomeUnknownValue = outcome;
+    this.submissionOutcomeUnknown = outcome;
     this.callbacks.requestUpdate();
   }
 
@@ -320,7 +316,7 @@ export class DraftSubmissionFlow {
         pendingPlacement: this.pendingPlacement,
         submitting: this.submittingValue,
         message: this.messageValue,
-        submissionOutcomeUnknown: this.submissionOutcomeUnknownValue,
+        submissionOutcomeUnknown: this.submissionOutcomeUnknown,
         pendingAttachmentReads: this.attachmentDraft.pendingReads,
         hasDraftAttachments: this.attachmentDraft.attachments.length > 0,
         hasCapabilityOverrides: this.capabilities.toolOverrides !== null,
@@ -366,7 +362,7 @@ export class DraftSubmissionFlow {
       (outcomeUnknown && this.submittingValue && !this.sessionStartup.interrupt()) ||
       this.sessionStartup.retireChangedOwner()
     ) {
-      this.submissionOutcomeUnknownValue = outcomeUnknown;
+      this.submissionOutcomeUnknown = outcomeUnknown;
     }
     this.submittingValue = false;
     this.callbacks.requestUpdate();
@@ -377,8 +373,8 @@ export class DraftSubmissionFlow {
     const preservePendingPlacement = Boolean(this.pendingPlacement.sessionKey);
     this.blockedSubmitGate = null;
     this.invalidate();
-    this.submissionOutcomeUnknownValue = preservePendingPlacement
-      ? (this.submissionOutcomeUnknownValue ?? "placement-interrupted")
+    this.submissionOutcomeUnknown = preservePendingPlacement
+      ? (this.submissionOutcomeUnknown ?? "placement-interrupted")
       : null;
     this.visibilityValue = "normal";
     this.capabilities.reset();
@@ -404,22 +400,21 @@ export class DraftSubmissionFlow {
 
   clearPendingPlacementRecovery() {
     this.pendingPlacement.clear();
-    this.submissionOutcomeUnknownValue = null;
+    this.submissionOutcomeUnknown = null;
     this.callbacks.requestUpdate();
   }
 
   releasePendingPlacementOwner() {
     this.pendingPlacement.reset();
-    this.submissionOutcomeUnknownValue = null;
+    this.submissionOutcomeUnknown = null;
     this.callbacks.requestUpdate();
   }
 
   restorePendingPlacementRecovery(gatewayUrl: string, recoveryScope: string) {
     const recovery = this.pendingPlacement.restore(gatewayUrl, recoveryScope);
-    if (!recovery) {
-      return;
+    if (recovery) {
+      this.applyRecoveryDraft(recovery);
     }
-    this.applyRecoveryDraft(recovery);
   }
 
   async submit(startup?: { params: SessionCreateParams; startedAt: number }) {
