+++ ui/src/pages/new-session/model-control.test.ts
@@ -215,7 +215,7 @@ describe("new-session model runtime", () => {
       true,
     );
     expect(loadingModelTrigger?.getAttribute("aria-label")).toBe("Chat model: Loading models…");
-    expect(loadingModelTrigger?.getAttribute("aria-disabled")).toBe("true");
+    expect(loadingModelTrigger?.getAttribute("aria-disabled")).toBe("false");
     expect(loadingSkeleton).not.toBeNull();
     expect(loadingSkeleton?.getAttribute("aria-hidden")).toBe("true");
     expect(loadingModelTrigger?.textContent).not.toContain("Loading models");
@@ -251,7 +251,11 @@ describe("new-session model runtime", () => {
     expect(loadingSkeleton?.getAttribute("aria-hidden")).toBe("true");
     expect(loadingModelTrigger?.textContent).not.toContain("Loading models");
     expect(loadingModelTrigger?.textContent).not.toContain("Default model");
-    expect(container.querySelector('[data-chat-thinking-select="true"]')).toBeNull();
+    const reservedThinkingControl = container.querySelector<HTMLElement>(
+      '[data-chat-thinking-select="true"]',
+    );
+    expect(reservedThinkingControl).not.toBeNull();
+    expect(reservedThinkingControl?.dataset.chatThinkingDisabled).toBe("true");
     expect(control.selected).toBe("");
     expect(control.thinkingLevel).toBe("");
